Directional lights well well now.
All directional lights are handled with a shared set of VPLs:
Base ray (plus oversampling) are cast from the camera, then backed off a bit.
Those hit points are checked against each directional light to see if they light it.
If so, we store the total energy, average color, position, and normal as a VPL