Added surface offset
Added a surface offset option to move the VPL away from the (average) surface normal.
Reorganized the export groups.
Changed the VPL attenuation to not flatten it quite so much.
If FauxGI is disabled don't keep updating the ambient term...it interfered with changing the WorldEnvironment settings inside the editor.
Free Virtual Directional Lights on _exit_tree().